+++ Source/02_MLAMotion_lib/MLAMotion/MLAJoint.h
@@ -23,11 +23,11 @@ public:
 	Joint& operator=(const Joint& joint);
 
 	void setJointName(const std::string& jointName);
-	void setPositions(const glm::dvec3& positions);
+	void setPositions(const glm::vec3& positions);
 	void setOrientations(const glm::quat& orientations);
 
 	const std::string& getJointName() const;
-	const glm::dvec3& getPositions() const;
+	const glm::vec3& getPositions() const;
 	const glm::quat& getOrientations() const;
 
 	void setParent(Joint* parent);
@@ -40,7 +40,7 @@ private:
 	std::string m_jointName;
 
 	// a vector with the initial offset
-	glm::dvec3 m_positions;
+	glm::vec3 m_positions;
 
 	// a quaternion holding the orientation for the frame
 	glm::quat m_orientations;
